+++ imap_tools/__init__.py
@@ -1,5 +1,6 @@
 # Lib author: Vladimir Kaukin <KaukinVK@ya.ru>
 # Project home page: https://github.com/ikvk/imap_tools
+# License: Apache-2.0
 
 from .query import AND, OR, NOT, Header, UidRange, A, O, N, H, U
 from .mailbox import BaseMailBox, MailBox, MailBoxUnencrypted
@@ -8,4 +9,4 @@ from .folder import MailBoxFolderManager
 from .consts import MailMessageFlags, MailBoxFolderStatusOptions
 from .errors import *
 
-__version__ = '0.46.0'
+__version__ = '0.47.0'