+++ imap_tools/query.py
@@ -3,6 +3,7 @@ import datetime
 import itertools
 import functools
 import collections
+from typing import Iterable, Optional, Dict, Any, List, Union, Sequence
 
 from .consts import SHORT_MONTH_NAMES
 from .utils import clean_uids, quote
@@ -24,7 +25,7 @@ class LogicOperator(collections.UserString):
         raise NotImplementedError
 
     @staticmethod
-    def prefix_join(operator: str, params: iter) -> str:
+    def prefix_join(operator: str, params: Iterable[str]) -> str:
         """Join params by prefix notation rules, enclose in parenthesis"""
         return '({})'.format(functools.reduce(lambda a, b: '{}{} {}'.format(operator, a, b), params))
 
@@ -50,12 +51,6 @@ class NOT(LogicOperator):
         return 'NOT {}'.format(self.prefix_join('', itertools.chain(self.converted_strings, self.converted_params)))
 
 
-# Short alias set:
-A = AND
-O = OR  # noqa
-N = NOT
-
-
 class Header:
     __slots__ = ('name', 'value')
 
@@ -71,9 +66,6 @@ class Header:
         return '{0.name}: {0.value}'.format(self)
 
 
-H = Header  # Short alias
-
-
 class UidRange:
     """
     * - represents the largest number in use.
@@ -84,7 +76,7 @@ class UidRange:
     """
     __slots__ = ('start', 'end')
 
-    def __init__(self, start: str, end: str or None = None):
+    def __init__(self, start: str, end: Optional[str] = None):
         self.start = str(start).strip()
         if not (self.start.isdigit() or self.start == '*'):
             raise TypeError('UidRange start arg must be str with digits or *')
@@ -99,9 +91,6 @@ class UidRange:
         return '{}{}'.format(self.start, ':{}'.format(self.end) if self.end else '')
 
 
-U = UidRange  # Short alias
-
-
 class ParamConverter:
     """Convert search params to IMAP format"""
 
@@ -111,10 +100,10 @@ class ParamConverter:
         'header', 'gmail_label',
     )
 
-    def __init__(self, params: dict):
+    def __init__(self, params: Dict[str, Any]):
         self.params = params
 
-    def _gen_values(self, key, value) -> iter:
+    def _gen_values(self, key: str, value: Any) -> Iterable[Any]:
         """Values generator"""
         # single value
         if key not in self.multi_key_allowed or isinstance(value, str):
@@ -128,7 +117,7 @@ class ParamConverter:
                 # single value
                 yield value
 
-    def convert(self) -> [str]:
+    def convert(self) -> List[str]:
         """
         :return: params in IMAP format
         """
@@ -147,37 +136,37 @@ class ParamConverter:
         return '{}-{}-{}'.format(value.day, SHORT_MONTH_NAMES[value.month - 1], value.year)
 
     @staticmethod
-    def cleaned_str(key, value) -> str:
+    def cleaned_str(key: str, value: str) -> str:
         if type(value) is not str:
             raise TypeError('"{}" expected str value, "{}" received'.format(key, type(value)))
         return str(value)
 
     @staticmethod
-    def cleaned_date(key, value) -> datetime.date:
+    def cleaned_date(key: str, value: datetime.date) -> datetime.date:
         if type(value) is not datetime.date:
             raise TypeError('"{}" expected datetime.date value, "{}" received'.format(key, type(value)))
         return value
 
     @staticmethod
-    def cleaned_bool(key, value) -> bool:
+    def cleaned_bool(key: str, value: bool) -> bool:
         if type(value) is not bool:
             raise TypeError('"{}" expected bool value, "{}" received'.format(key, type(value)))
         return bool(value)
 
     @staticmethod
-    def cleaned_true(key, value) -> True:
+    def cleaned_true(key: str, value: bool) -> True:
         if value is not True:
             raise TypeError('"{}" expected "True", "{}" received'.format(key, type(value)))
         return True
 
     @staticmethod
-    def cleaned_uint(key, value) -> int:
+    def cleaned_uint(key: str, value: int) -> int:
         if type(value) is not int or int(value) < 0:
             raise TypeError('"{}" expected int value >= 0, "{}" received'.format(key, type(value)))
         return int(value)
 
     @staticmethod
-    def cleaned_uid(key, value) -> str:
+    def cleaned_uid(key: str, value: Union[str, Sequence[str], UidRange]) -> str:
         # range
         if isinstance(value, UidRange):
             return str(value)
@@ -188,118 +177,118 @@ class ParamConverter:
             raise TypeError('{} parse error: {}'.format(key, str(e)))
 
     @staticmethod
-    def cleaned_header(key, value) -> H:
-        if not isinstance(value, H):
+    def cleaned_header(key: str, value: Header) -> Header:
+        if not isinstance(value, Header):
             raise TypeError('"{}" expected Header (H) value, "{}" received'.format(key, type(value)))
         return value
 
-    def convert_answered(self, key, value):
+    def convert_answered(self, key, value) -> str:
         """Messages [with/without] the Answered flag set. (ANSWERED, UNANSWERED)"""
         return 'ANSWERED' if self.cleaned_bool(key, value) else 'UNANSWERED'
 
-    def convert_seen(self, key, value):
+    def convert_seen(self, key, value) -> str:
         """Messages that [have/do not have] the Seen flag set. (SEEN, UNSEEN)"""
         return 'SEEN' if self.cleaned_bool(key, value) else 'UNSEEN'
 
-    def convert_flagged(self, key, value):
+    def convert_flagged(self, key, value) -> str:
         """Messages [with/without] the Flagged flag set. (FLAGGED, UNFLAGGED)"""
         return 'FLAGGED' if self.cleaned_bool(key, value) else 'UNFLAGGED'
 
-    def convert_draft(self, key, value):
+    def convert_draft(self, key, value) -> str:
         """Messages that [have/do not have] the Draft flag set. (DRAFT, UNDRAFT)"""
         return 'DRAFT' if self.cleaned_bool(key, value) else 'UNDRAFT'
 
-    def convert_deleted(self, key, value):
+    def convert_deleted(self, key, value) -> str:
         """Messages that [have/do not have] the Deleted flag set. (DELETED, UNDELETED)"""
         return 'DELETED' if self.cleaned_bool(key, value) else 'UNDELETED'
 
-    def convert_keyword(self, key, value):
+    def convert_keyword(self, key, value) -> str:
         """Messages with the specified keyword flag set. (KEYWORD)"""
         return 'KEYWORD {}'.format(self.cleaned_str(key, value))
 
-    def convert_no_keyword(self, key, value):
+    def convert_no_keyword(self, key, value) -> str:
         """Messages that do not have the specified keyword flag set. (UNKEYWORD)"""
         return 'UNKEYWORD {}'.format(self.cleaned_str(key, value))
 
-    def convert_from_(self, key, value):
+    def convert_from_(self, key, value) -> str:
         """Messages that contain the specified string in the envelope structure's FROM field."""
         return 'FROM {}'.format(quote(self.cleaned_str(key, value)))
 
-    def convert_to(self, key, value):
+    def convert_to(self, key, value) -> str:
         """Messages that contain the specified string in the envelope structure's TO field."""
         return 'TO {}'.format(quote(self.cleaned_str(key, value)))
 
-    def convert_subject(self, key, value):
+    def convert_subject(self, key, value) -> str:
         """Messages that contain the specified string in the envelope structure's SUBJECT field."""
         return 'SUBJECT {}'.format(quote(self.cleaned_str(key, value)))
 
-    def convert_body(self, key, value):
+    def convert_body(self, key, value) -> str:
         """Messages that contain the specified string in the body of the message."""
         return 'BODY {}'.format(quote(self.cleaned_str(key, value)))
 
-    def convert_text(self, key, value):
+    def convert_text(self, key, value) -> str:
         """Messages that contain the specified string in the header or body of the message."""
         return 'TEXT {}'.format(quote(self.cleaned_str(key, value)))
 
-    def convert_bcc(self, key, value):
+    def convert_bcc(self, key, value) -> str:
         """Messages that contain the specified string in the envelope structure's BCC field."""
         return 'BCC {}'.format(quote(self.cleaned_str(key, value)))
 
-    def convert_cc(self, key, value):
+    def convert_cc(self, key, value) -> str:
         """Messages that contain the specified string in the envelope structure's CC field."""
         return 'CC {}'.format(quote(self.cleaned_str(key, value)))
 
-    def convert_date(self, key, value):
+    def convert_date(self, key, value) -> str:
         """
         Messages whose internal date (disregarding time and timezone)
         is within the specified date. (ON)
         """
         return 'ON {}'.format(self.format_date(self.cleaned_date(key, value)))
 
-    def convert_date_gte(self, key, value):
+    def convert_date_gte(self, key, value) -> str:
         """
         Messages whose internal date (disregarding time and timezone)
         is within or later than the specified date. (SINCE)
         """
         return 'SINCE {}'.format(self.format_date(self.cleaned_date(key, value)))
 
-    def convert_date_lt(self, key, value):
+    def convert_date_lt(self, key, value) -> str:
         """
         Messages whose internal date (disregarding time and timezone)
         is earlier than the specified date. (BEFORE)
         """
         return 'BEFORE {}'.format(self.format_date(self.cleaned_date(key, value)))
 
-    def convert_sent_date(self, key, value):
+    def convert_sent_date(self, key, value) -> str:
         """
         Messages whose [RFC-2822] Date: header (disregarding time and timezone)
         is within the specified date. (SENTON)
         """
         return 'SENTON {}'.format(self.format_date(self.cleaned_date(key, value)))
 
-    def convert_sent_date_gte(self, key, value):
+    def convert_sent_date_gte(self, key, value) -> str:
         """
         Messages whose [RFC-2822] Date: header (disregarding time and timezone)
         is within or later than the specified date. (SENTSINCE)
         """
         return 'SENTSINCE {}'.format(self.format_date(self.cleaned_date(key, value)))
 
-    def convert_sent_date_lt(self, key, value):
+    def convert_sent_date_lt(self, key, value) -> str:
         """
         Messages whose [RFC-2822] Date: header (disregarding time and timezone)
         is earlier than the specified date. (SENTBEFORE)
         """
         return 'SENTBEFORE {}'.format(self.format_date(self.cleaned_date(key, value)))
 
-    def convert_size_gt(self, key, value):
+    def convert_size_gt(self, key, value) -> str:
         """Messages with an [RFC-2822] size larger than the specified number of octets. (LARGER)"""
         return 'LARGER {}'.format(self.cleaned_uint(key, value))
 
-    def convert_size_lt(self, key, value):
+    def convert_size_lt(self, key, value) -> str:
         """Messages with an [RFC-2822] size smaller than the specified number of octets. (SMALLER)"""
         return 'SMALLER {}'.format(self.cleaned_uint(key, value))
 
-    def convert_new(self, key, value):
+    def convert_new(self, key, value) -> str:
         """
         Messages that have the Recent flag set but not the Seen flag.
         This is functionally equivalent to "(RECENT UNSEEN)".
@@ -307,7 +296,7 @@ class ParamConverter:
         self.cleaned_true(key, value)
         return 'NEW'
 
-    def convert_old(self, key, value):
+    def convert_old(self, key, value) -> str:
         """
         Messages that do not have the Recent flag set.
         This is functionally equivalent to "NOT RECENT" (as opposed to "NOT NEW").
@@ -315,17 +304,17 @@ class ParamConverter:
         self.cleaned_true(key, value)
         return 'OLD'
 
-    def convert_recent(self, key, value):
+    def convert_recent(self, key, value) -> str:
         """Messages that have the Recent flag set."""
         self.cleaned_true(key, value)
         return 'RECENT'
 
-    def convert_all(self, key, value):
+    def convert_all(self, key, value) -> str:
         """All messages in the mailbox; the default initial key for ANDing."""
         self.cleaned_true(key, value)
         return 'ALL'
 
-    def convert_header(self, key, value):
+    def convert_header(self, key, value) -> str:
         """
         Messages that have a header with the specified field-name (as defined in [RFC-2822])
         and that contains the specified string in the text of the header (what comes after the colon).
@@ -334,9 +323,17 @@ class ParamConverter:
         """
         return 'HEADER {0.name} {0.value}'.format(self.cleaned_header(key, value))
 
-    def convert_uid(self, key, value):
+    def convert_uid(self, key, value) -> str:
         """Messages with unique identifiers corresponding to the specified unique identifier set."""
         return 'UID {}'.format(self.cleaned_uid(key, value))
 
-    def convert_gmail_label(self, key, value):
+    def convert_gmail_label(self, key, value) -> str:
         return 'X-GM-LABELS {}'.format(quote(self.cleaned_str(key, value)))
+
+
+# Short alias set:
+A = AND
+O = OR  # noqa
+N = NOT
+H = Header
+U = UidRange